Miguel - Small impovement in Stat

+++ artemis-metrics/src/main/java/com/spinn3r/artemis/metrics/Stat.java
@@ -61,7 +61,7 @@ public class Stat<T> {
      * @param value
      * @param delta
      */
-    public void incr( T value, int delta ) {
+    public void incr( T value, long delta ) {
 
         if ( delta == 0 ) {
